Remove unnecessary method call to skip snapshot fetch during release workflow (#47)
This PR removes the maven_dependencies_update_after_release method call, which was unnecessary and caused Maven to fetch SNAPSHOT dependencies during the release process.

The method was originally used to update the next snapshot version after the project release. However, the missing part of the workflow—pushing the release branch—was never implemented. Additionally, updating the snapshot version triggered Maven to fetch SNAPSHOT dependencies, which we don't need. As a result, this method call has been removed as a workaround to skip the unnecessary snapshot fetch, streamlining the release process.